requirements.txt
fastapi
uvicorn
openai
spacy
python-dotenv
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

